pyproject.toml
[project]
name = "arxiv-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cohere>=5.11.0",
    "feedparser>=6.0.11",
    "langchain-core>=0.3.10",
    "langchain-openai>=0.2.2",
    "langgraph-checkpoint>=2.0.1",
    "langgraph>=0.2.35",
    "pydantic-settings>=2.5.2",
    "streamlit>=1.39.0",
]
